implemented look-ahead candidate selection (sorta)
in a way, I had already accidentaly implemented a less efficient version of this with the MRV heuristic: my method that returned the first cell with the fewest candidates allowed for cells to have zero candidates. Therefore the algo speed I gained is only equal to the number of times a constraint propagation caused a candidateless cell, because now I catch these cases before taking that step, instead of taking the step and then backtracking immediately.